ZFP internal definitions moved to zfp-private.h
