+++ plugins/codecs/zfp/zfp-private.h
@@ -14,6 +14,9 @@
 #define ZFP_PRIVATE_H
 #include "context.h"
 
+#define ZFP_MAX_DIM 4
+#define ZFP_CELL_SHAPE 4
+
 #if defined (__cplusplus)
 extern "C" {
 #endif
@@ -113,6 +116,26 @@ static int32_t deserialize_meta(uint8_t *smeta, uint32_t smeta_len, int8_t *ndim
     }
     uint32_t slen = (uint32_t)(pmeta - smeta);
     return 0;
+}
+
+static void index_unidim_to_multidim(int8_t ndim, int64_t *shape, int64_t i, int64_t *index) {
+    int64_t strides[ZFP_MAX_DIM];
+    strides[ndim - 1] = 1;
+    for (int j = ndim - 2; j >= 0; --j) {
+        strides[j] = shape[j + 1] * strides[j + 1];
+    }
+
+    index[0] = i / strides[0];
+    for (int j = 1; j < ndim; ++j) {
+        index[j] = (i % strides[j - 1]) / strides[j];
+    }
+}
+
+static void index_multidim_to_unidim(int64_t *index, int8_t ndim, int64_t *strides, int64_t *i) {
+    *i = 0;
+    for (int j = 0; j < ndim; ++j) {
+        *i += index[j] * strides[j];
+    }
 }
 
 #if defined (__cplusplus)
